--- interpreter/guihelper.hpp
@@ -1,11 +1,0 @@
-//
-// Created by OA82HX on 08-09-2022.
-//
-
-#ifndef MINIMC_GUIHELPER_HPP
-#define MINIMC_GUIHELPER_HPP
-
-MiniMC::Model::Edge* promptForEdge(MiniMC::CPA::AnalysisState state);
-void printState(MiniMC::CPA::AnalysisState state);
-
-#endif // MINIMC_GUIHELPER_HPP
